+++ src/pages/admin/stats.tsx
@@ -1,15 +1,19 @@
-import AdminPage from '@/components/Admin/AdminPage/adminPage'
-import Summary from '@/components/Admin/Stats/summary'
-
-const dummyUsername = 'Admin'
+import VoluteerStatsTable from '@/components/Admin/volunteerstatstable'
+import Header from '@/components/Header'
+import TopNav from '@/components/TopNav'
 
 const Stats = () => {
   return (
-    <AdminPage
-      title='Admin Statistics'
-      heading={`Welcome ${dummyUsername}!`}
-      content={<Summary />}
-    />
+    <>
+      <Header pageTitle='Admin' />
+      <TopNav />
+      <main>
+        <h1 className='my-8 mx-auto text-center text-4xl font-bold text-primary-dark'>
+          Admin Summary Dashboard
+        </h1>
+        <VoluteerStatsTable />
+      </main>
+    </>
   )
 }
